requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# File Upload
python-multipart==0.0.6
aiofiles==23.2.1

# Templating (for serving HTML)
jinja2==3.1.3

# Development
pytest==7.4.3
httpx==0.26.0
